requirements.txt
poetry==1.8.3
boto3==1.35.28
openai==1.48.0
pyyaml==6.0.2
black==24.8.0